requirements.txt
streamlit
openai
jsonschema
logging
pytest
streamlit
google-generativeai
python-dotenv